Revert to not using full parent paths, because we want to be able to move notes.
This incurs tombstone problem issues.  There is a large chance of
generating orphan records.    In particular, any comments on a post that
are newer than the deletion time itself can't safely be deleted in a
protocol that allows for reanimation...

99.9% of the time this should be perfectly fine.  The rest will need an
orphan nodes browser to really be safe from accidentally revealling
stuff you did not want to.